v_beta/robot_arm_simulator/requirements.txt
# Core dependencies
numpy>=1.20.0
pybullet>=3.2.0
PyYAML>=6.0
matplotlib>=3.4.0

# Development dependencies
pytest>=6.0
pytest-cov>=2.0
black>=21.0
isort>=5.0
mypy>=0.9
sphinx>=4.0
sphinx-rtd-theme>=0.5.0
